+++ source/index.ts
@@ -1,68 +1,13 @@
-/* eslint no-console:0 */
-'use strict'
-
 // Imports
 import { resolve } from 'path'
 import semver from 'semver'
-import { errtion, stringify, simplifyRange } from './util.js'
+import { errtion, Errtion } from './util.js'
+import { versions as processVersions } from 'process'
+import { readFileSync } from 'fs'
 
-// As Errlop uses Editions, we should use a specific Errlop edition
-// As otherwise, the circular reference may fail on some machines
-// https://github.com/bevry/errlop/issues/2
-import Errlop from 'errlop'
-
-/**
- * The current node version that we are operating within.
- * It is compared in {@link requireEdition} against {@link Edition.engines}.
- */
-const NODE_VERSION: string = process.versions.node
-
-/**
- * Set the environment variable `EDITIONS_VERBOSE` to output debugging information to stderr on how editions selected the edition it did.
- * Values of `yes` and `true` are supported.
- * @example env EDITIONS_VERBOSE=true node mypackage/index.js
- */
-type EDITIONS_VERBOSE = string | boolean | undefined
-
-/**
- * Whether or not {@link EDITIONS_VERBOSE} is enabled.
- * @type {boolean}
- * @private
- */
-const VERBOSE: boolean =
-	(process.env.EDITIONS_VERBOSE as EDITIONS_VERBOSE) === true ||
-	(process.env.EDITIONS_VERBOSE as EDITIONS_VERBOSE) === 'yes' ||
-	(process.env.EDITIONS_VERBOSE as EDITIONS_VERBOSE) === 'true' ||
-	false
-
-/**
- * Set the environment variable `EDITIONS_TAG_BLACKLIST` to the tags you wish to blacklist, and editions will skip editions that contain them.
- * For backwards compatibility `EDITIONS_SYNTAX_BLACKLIST` is also supported.
- * It is compared in {@link requireEdition} against {@link Edition.tags}.
- * The value of this is stored locally in the {@link BLACKLIST} cache.
- * @example env EDITIONS_TAG_BLACKLIST=esnext,typescript,coffeescript node mypackage/index.js
- */
-type EDITIONS_TAG_BLACKLIST = string
-
-/**
- * A list of the blacklisted tags.
- * Data imported from {@link EDITIONS_TAG_BLACKLIST}.
- */
-const BLACKLIST: string[] =
-	(process.env.EDITIONS_TAG_BLACKLIST as EDITIONS_TAG_BLACKLIST)?.split(
-		/[, ]+/g
-	) ||
-	(process.env.EDITIONS_SYNTAX_BLACKLIST as EDITIONS_TAG_BLACKLIST)?.split(
-		/[, ]+/g
-	)
-
-/**
- * A mapping of blacklisted tags to their reasons.
- * Keys are the tags.
- * Values are the error instances that contain the reasoning for why/how that tag is/became blacklisted.
- * Data imported from {@link EDITIONS_TAG_BLACKLIST}.
- */
-const blacklist: { [tag: string]: Error } = {}
+export type Range = string | boolean
+export type Engines = false | { [engine: string]: Range }
+export type Versions = { [engine: string]: string }
 
 /**
  * Edition entries must conform to the following specification.
@@ -92,7 +37,7 @@ export interface Edition {
 	 * "esnext source code with require for modules"
 	 * ```
 	 */
-	description: String
+	description: string
 	/**
 	 * The location to where this directory is located. It should be a relative path from the `package.json` file.
 	 * @example
@@ -108,9 +53,9 @@ export interface Edition {
 	 * "index.js"
 	 * ```
 	 */
-	entry?: string
+	entry: string
 	/**
-	 * Any keywords you wish to associate to the edition. Useful for various ecosystem tooling, such as automatic ESNext lint configuration if the `esnext` tag is present in the source edition tags. Consumers also make use of this via {@link EDITIONS_TAG_BLACKLIST} for preventing loading editions that contain a blacklisted tag. Previously this field was named `syntaxes`.
+	 * Any keywords you wish to associate to the edition. Useful for various ecosystem tooling, such as automatic ESNext lint configuration if the `esnext` tag is present in the source edition tags.
 	 * @example
 	 * ``` json
 	 * ["javascript", "esnext", "require"]
@@ -118,86 +63,82 @@ export interface Edition {
 	 */
 	tags?: string[]
 	/**
-	 * Alias of {@link Edition['tags']} kept for backwards compatibility with legacy edition definitions
-	 * @ignore
-	 */
-	syntaxes?: Edition['tags']
-	/**
-	 * This field is used to specific which Node.js and Browser environments this edition supports. If `false` this edition does not support either. If `node` is a string, it should be a semver range of node.js versions that the edition targets. If `browsers` is a string, it should be a [browserlist](https://github.com/browserslist/browserslist) value of the specific browser values the edition targets. If `node` or `browsers` is true, it indicates that this edition is compatible with those environments.
+	 * This field is used to specific which environments this edition supports.
+	 * If `false` this edition does not any environment.
+	 * If `deno` is a string, it should be a semver range of Deno versions that the edition targets.
+	 * If `node` is a string, it should be a semver range of Node.js versions that the edition targets.
+	 * If `browsers` is a string, it should be a [browserlist](https://github.com/browserslist/browserslist) value of the specific browser values the edition targets. If multiple engines are truthy, it indicates that this edition is compatible with those multiple environments.
 	 * @example
 	 * ``` json
 	 * {
+	 *   "deno": ">=1",
 	 *   "node": ">=6",
 	 *   "browsers": "defaults"
 	 * }
 	 * ```
 	 */
-	engines?: false | { [engine: string]: string | boolean }
+	engines: Engines
 }
 
-/** These are the various options that you can use to customise the behaviour of certain methods. */
-export interface EditionOptions {
-	/** The require method of the calling module, used to ensure require paths remain correct. */
-	require: Function
-	/** If provided, this is used for debugging. */
-	packagePath?: string
-	/** If provided, any error loading an edition will be logged. By default, errors are only logged if all editions failed. If not provided, process.env.EDITIONS_VERBOSE is used. */
-	verbose?: boolean
-	/** If `true`, then only exact version matches will be loaded. If `false`, then likely matches using {@link simplifyRange} will be evaluated, with a fallback to the last. If missing, then `true` is attempted first and if no result, then `false` is attempted. */
-	strict?: boolean
-	/** If provided, this will be the cwd for entries. */
-	cwd?: string
-	/** If provided, should be a relative path to the entry point of the edition. */
-	entry?: string
-	/** If provided, should be the name of the package that we are loading the editions for. */
-	package?: string
-	/** If not provided, will use process.stderr instead. It is the stream that verbose errors are logged to. */
-	stderr?: NodeJS.WritableStream
+/** Editions should be ordered from most preferable first to least desirable last. The source edition should always be first, proceeded by compiled editions. */
+export type Editions = Array<Edition>
+
+export interface PathOptions {
+	/** If provided, this edition entry is used instead of the default entry. */
+	entry: string
+
+	/** If provided, edition loading will be resolved against this. */
+	cwd: string
 }
 
-// Create the mapping of blacklisted tags and their reasonings
-if (BLACKLIST) {
-	for (let i = 0; i < BLACKLIST.length; ++i) {
-		const tag = BLACKLIST[i].trim().toLowerCase()
-		blacklist[tag] = errtion({
-			message: `The EDITIONS_TAG_BLACKLIST (aka EDITIONS_SYNTAX_BLACKLIST) environment variable blacklisted the tag [${tag}]`,
-			code: 'blacklisted-tag',
-		})
-	}
+export interface LoaderOptions extends Partial<PathOptions> {
+	/**
+	 * The method that will load the entry of the edition.
+	 * For CJS files this should be set to the `require` method.
+	 * For MJS files this should be set to `(path: string) => import(path)`.
+	 */
+	loader: <T>(this: Edition, path: string) => T
 }
 
-// Blacklist the tag 'esnext' if our node version is below 0.12
-if (semver.satisfies(NODE_VERSION, '<0.12')) {
-	blacklist.esnext = new Error(
-		'The esnext tag is skipped on early node versions as attempting to use esnext features will output debugging information on these node versions'
-	)
+export interface RangeOptions {
+	/** If `true`, then ranges such as `x || y` are changed to `>=x`. */
+	broadenRange?: boolean
 }
+
+export interface VersionOptions extends RangeOptions {
+	/** The versions of our current environment. */
+	versions: Versions
+}
+
+export interface SolicitOptions extends LoaderOptions, VersionOptions {}
 
 /**
- * Attempt to load a specific {@link Edition}.
+ * Load the {@link Edition} with the loader.
  * @returns The result of the loaded edition.
- * @throws An error if the edition failed to load.
+ * @throws If failed to load, an error is thrown with the reason.
  */
-export function loadEdition(edition: Edition, opts: EditionOptions): any {
+export function loadEdition<T>(edition: Edition, opts: LoaderOptions): T {
 	const entry = resolve(
 		opts.cwd || '',
 		edition.directory,
 		opts.entry || edition.entry || ''
 	)
-	if (opts.require == null) {
+	if (opts.loader == null) {
 		throw errtion({
-			message: `Skipped edition [${edition.description}] as opts.require was not provided, this is probably due to a testing misconfiguration.`,
-			code: 'unsupported-edition-require',
+			message: `Could not load the edition [${edition.description}] as no loader was specified. This is probably due to a testing misconfiguration.`,
+			code: 'editions-autoloader-loader-missing',
+			level: 'fatal',
 		})
 	}
 	try {
-		return opts.require(entry)
+		return opts.loader.call(edition, entry) as T
 	} catch (loadError) {
 		// Note the error with more details
 		throw errtion(
 			{
-				message: `Skipped edition [${edition.description}] at entry [${entry}] because it failed to load`,
-				code: 'unsupported-edition-tried',
+				message: `Failed to load the entry [${entry}] of edition [${edition.description}].`,
+				code: 'editions-autoloader-loader-failed',
+				level: 'fatal',
 			},
 			loadError
 		)
@@ -205,13 +146,11 @@ export function loadEdition(edition: Edition, opts: EditionOptions): any {
 }
 
 /**
- * Attempt to require an {@link Edition}, based on its compatibility with the current environment, such as {@link NODE_VERSION} and {@link EDITIONS_TAG_BLACKLIST} compatibility.
- * If compatibility is established with the environment, it will load the edition using {@link loadEdition}.
- * @returns The result of the loaded edition
- * @throws An error if the edition failed to load
+ * Verify the {@link Edition} has all the required properties.
+ * @returns if valid
+ * @throws if invalid
  */
-export function requireEdition(edition: Edition, opts: EditionOptions): any {
-	// Verify the edition is valid
+export function isValidEdition(edition: Edition): true {
 	if (
 		!edition.description ||
 		!edition.directory ||
@@ -219,217 +158,285 @@ export function requireEdition(edition: Edition, opts: EditionOptions): any {
 		edition.engines == null
 	) {
 		throw errtion({
-			message: `Each edition must have its [description, directory, entry, engines] fields defined, yet all it had was [${Object.keys(
+			message: `An edition must have its [description, directory, entry, engines] fields defined, yet all this edition defined were [${Object.keys(
 				edition
 			).join(', ')}]`,
-			code: 'unsupported-edition-malformed',
+			code: 'editions-autoloader-invalid-edition',
 			level: 'fatal',
 		})
 	}
 
-	// Handle strict omission
-	if (opts.strict == null) {
-		try {
-			return requireEdition(edition, { ...opts, strict: true })
-		} catch (err) {
-			return requireEdition(edition, { ...opts, strict: false })
-		}
-	}
-
-	// Verify tag support
-	// Convert tags into a sorted lowercase string
-	const tags = (edition.tags || edition.syntaxes || [])
-		.map((i) => i.toLowerCase())
-		.sort()
-	for (let index = 0; index < tags.length; index++) {
-		const tag = tags[index]
-		const blacklisted = blacklist[tag]
-		if (blacklisted) {
-			throw errtion(
-				{
-					message: `Skipping edition [${edition.description}] because it contained a blacklisted tag [${tag}]`,
-					code: 'unsupported-edition-backlisted-tag',
-				},
-				blacklisted
-			)
-		}
-	}
-
-	// Verify engine support
-	if (edition.engines === false) {
-		throw errtion({
-			message: `Skipping edition [${edition.description}] because its engines field was false`,
-			code: 'unsupported-edition-engine',
-		})
-	}
-	if (!edition.engines.node) {
-		throw errtion({
-			message: `Skipping edition [${edition.description}] because its .engines.node field was falsey`,
-			code: 'unsupported-edition-engines-node',
-		})
-	}
-	if (opts.strict) {
-		if (edition.engines.node === true) {
-			throw errtion({
-				message: `Skipping edition [${edition.description}] because its .engines.node field was true yet we are in strict mode`,
-				code: 'unsupported-edition-engines-node-version-true',
-			})
-		} else if (semver.satisfies(NODE_VERSION, edition.engines.node) === false) {
-			throw errtion({
-				message: `Skipping edition [${
-					edition.description
-				}] because our current node version [${NODE_VERSION}] is not supported by its specific range [${stringify(
-					edition.engines.node
-				)}]`,
-				code: 'unsupported-edition-engines-node-version-specific',
-			})
-		}
-	} else if (edition.engines.node !== true) {
-		const simplifiedRange = simplifyRange(edition.engines.node)
-		if (semver.satisfies(NODE_VERSION, simplifiedRange) === false) {
-			throw errtion({
-				message: `Skipping edition [${
-					edition.description
-				}] because our current node version [${NODE_VERSION}] is not supported by its simplified range [${stringify(
-					simplifiedRange
-				)}]`,
-				code: 'unsupported-edition-engines-node-version-simplified',
-			})
-		}
-	}
-
-	// Load the edition
-	return loadEdition(edition, opts)
+	// valid
+	return true
 }
 
 /**
- * Cycles through a list of editions, returning the require result of the first suitable {@link Edition} that it was able to load.
- * Editions should be ordered from most preferable first, to least desirable last.
- * Providing the editions configuration is valid, individual edition handling is forwarded to {@link requireEdition}.
- * @returns The result of the loaded edition.
- * @throws An error if a suitable edition was unable to be resolved.
+ * Is this {@link Edition} suitable for these versions?
+ * @returns if compatible
+ * @throws if incompatible
  */
-export function requireEditions(
-	editions: Edition[],
-	opts: EditionOptions
-): any {
+export function isCompatibleVersion(
+	range: Range,
+	version: string,
+	opts: RangeOptions
+): true {
+	// prepare
+	const { broadenRange } = opts
+
+	if (!version)
+		throw errtion({
+			message: `No version was specified to compare the range [${range}] against`,
+			code: 'editions-autoloader-engine-version-missing',
+			level: 'fatal',
+		})
+
+	if (range == null || range === '')
+		throw errtion({
+			message: `The edition range was not specified, so unable to compare against the version [${version}]`,
+			code: 'editions-autoloader-engine-range-missing',
+		})
+
+	if (range === false)
+		throw errtion({
+			message: `The edition range does not support this engine`,
+			code: 'editions-autoloader-engine-unsupported',
+		})
+
+	if (range === true) return true
+
+	// semver compare
+	if (semver.satisfies(version, range)) {
+		return true
+	} else if (broadenRange === true) {
+		// broaden range
+		const index = range.indexOf('||')
+		if (index !== -1) {
+			const broadenedRange = range.substr(index)
+			// broadened range attempt
+			if (semver.satisfies(version, broadenedRange)) return true
+			// fail broadened range
+			throw errtion({
+				message: `The edition range [${range}] does not support this engine version [${version}], even when broadened to [${broadenedRange}]`,
+				code: 'editions-autoloader-engine-incompatible-broadened',
+			})
+		}
+		// give up
+		throw errtion({
+			message: `The edition range [${range}] does not support this engine version [${version}] and could not be broadened`,
+			code: 'editions-autoloader-engine-incompatible-nobroad',
+		})
+	} else {
+		// give up
+		throw errtion({
+			message: `The edition range [${range}] does not support this engine version [${version}]`,
+			code: 'editions-autoloader-engine-incompatible-original',
+		})
+	}
+
+	// return never
+}
+
+/**
+ * Checks that the provided engines are compatible against the provided versions.
+ * @returns if compatible
+ * @throws if incompatible
+ */
+export function isCompatibleEngines(
+	engines: Engines,
+	opts: VersionOptions
+): true {
+	// PRepare
+	const { versions } = opts
+
+	// Check engines exist
+	if (!engines) {
+		throw errtion({
+			message: `The edition had no engines to compare against the environment`,
+			code: 'editions-autoloader-invalid-engines',
+		})
+	}
+
+	// Check versions exist
+	if (!versions) {
+		throw errtion({
+			message: `No versions were supplied to compare the engines against`,
+			code: 'editions-autoloader-invalid-versions',
+			level: 'fatal',
+		})
+	}
+
+	// Check each version
+	let compatible = false
+	for (const key in engines) {
+		if (engines.hasOwnProperty(key)) {
+			// deno's std/node/process provides both `deno` and `node` keys
+			// so we don't won't to compare node when it is actually deno
+			if (key === 'node' && versions.deno) continue
+			// prepare
+			const engine = engines[key]
+			const version = versions[key]
+			// skip for engines this edition does not care about
+			if (version == null) continue
+			// check compatibility against all the provided engines it does care about
+			try {
+				isCompatibleVersion(engine, version, opts)
+				compatible = true
+				// if any incompatibility, it is thrown, so no need to set to false
+			} catch (rangeError) {
+				throw errtion(
+					{
+						message: `The engine [${key}] range of [${engine}] was not compatible against version [${version}].`,
+						code: 'editions-autoloader-engine-error',
+					},
+					rangeError
+				)
+			}
+		}
+	}
+
+	// if there were no matching engines, then throw
+	if (!compatible) {
+		throw errtion({
+			message: `There were no supported engines in which this environment provides.`,
+			code: 'editions-autoloader-engine-mismatch',
+		})
+	}
+
+	// valid
+	return true
+}
+
+/**
+ * Checks thaat the {@link Edition} is compatible against the provided versions.
+ * @returns if compatible
+ * @throws if incompatible
+ */
+export function isCompatibleEdition(
+	edition: Edition,
+	opts: VersionOptions
+): true {
+	try {
+		return isCompatibleEngines(edition.engines, opts)
+	} catch (compatibleError) {
+		throw errtion(
+			{
+				message: `The edition [${edition.description}] is not compatible with this environment.`,
+				code: 'editions-autoloader-edition-incompatible',
+			},
+			compatibleError
+		)
+	}
+}
+
+/**
+ * Determine which edition should be loaded.
+ * If {@link VersionOptions.broadenRange} is unspecified (the default behaviour), then we attempt to determine a suitable edition without broadening the range, and if that fails, then we try again with the range broadened.
+ * @returns any suitable editions
+ * @throws if no suitable editions
+ */
+export function determineEdition(
+	editions: Editions,
+	opts: VersionOptions
+): Edition {
+	// Prepare
+	const { broadenRange } = opts
+
 	// Check
 	if (!editions || editions.length === 0) {
-		if (opts.packagePath) {
-			throw errtion({
-				message: `There were no editions specified for package [${opts.packagePath}]`,
-				code: 'unsupported-editions-missing',
-			})
-		} else {
-			throw errtion({
-				message: 'There were no editions specified',
-				code: 'unsupported-editions-missing',
-			})
-		}
+		throw errtion({
+			message: 'No editions were specified.',
+			code: 'editions-autoloader-editions-missing',
+		})
 	}
-
-	// Handle strict omission
-	if (opts.strict == null) {
-		try {
-			return requireEditions(editions, { ...opts, strict: true })
-		} catch (err) {
-			return requireEditions(editions, { ...opts, strict: false })
-		}
-	}
-
-	// Whether or not we should be verbose
-	const verbose = opts.verbose == null ? VERBOSE : opts.verbose
-
-	// Capture the load result, the last error, and the fallback option
-	let result,
-		loaded = false,
-		editionsError = null,
-		fallbackEdition = null
 
 	// Cycle through the editions determing the above
+	let failure: Errtion | null = null
 	for (let i = 0; i < editions.length; ++i) {
 		const edition = editions[i]
 		try {
-			result = requireEdition(edition, opts)
-			loaded = true
-			break
-		} catch (editionError) {
-			if (editionError.level === 'fatal') {
-				editionsError = editionError
-				break
-			} else if (editionsError) {
-				editionsError = errtion(editionsError, editionError)
+			isValidEdition(edition)
+			isCompatibleEdition(edition, opts)
+			// return the edition if it is successful
+			return edition
+		} catch (error) {
+			if (error.level === 'fatal') {
+				throw errtion(
+					{
+						message: `Unable to determine a suitable edition due to failure.`,
+						code: 'editions-autoloader-fatal',
+						level: 'fatal',
+					},
+					error
+				)
+			} else if (failure) {
+				failure = errtion(error, failure)
 			} else {
-				editionsError = editionError
-			}
-			// make the fallback edition one that we don't bother loading due to its engines
-			// also: don't assume that .code is accessible, as it may not be, even if it should be, due to the way different environments behave
-			if (
-				String(editionError.code || '').indexOf(
-					'unsupported-edition-engines-node-version'
-				) === 0
-			) {
-				fallbackEdition = edition
+				failure = error
 			}
 		}
 	}
 
-	// if no edition was suitable for our environment, then try the fallback if it exists
-	// that is to say, ignore its engines.node
-	if (opts.strict === false && loaded === false && fallbackEdition) {
-		try {
-			result = loadEdition(fallbackEdition, opts)
-			loaded = true
-		} catch (editionError) {
-			editionsError = new Errlop(editionError, editionsError)
-		}
+	//
+	if (failure) {
+		// try broadened
+		if (broadenRange == null)
+			try {
+				return determineEdition(editions, { ...opts, broadenRange: true })
+			} catch (error) {
+				throw errtion(
+					{
+						message: `Unable to determine a suitable edition, even after broadening.`,
+						code: 'editions-autoloader-none-broadened',
+					},
+					error
+				)
+			}
+
+		// fail
+		throw errtion(
+			{
+				message: `Unable to determine a suitable edition, as none were suitable.`,
+				code: 'editions-autoloader-none-suitable',
+			},
+			failure
+		)
 	}
 
-	// if we were able to load something, then provide it
-	if (loaded) {
-		// make note of any errors if desired
-		if (editionsError && verbose) {
-			const stderr = opts.stderr || process.stderr
-			stderr.write(editionsError.stack + '\n')
-		}
-		return result
-	}
-	// otherwise, provide the error
-	else if (editionsError) {
-		if (opts.packagePath) {
-			throw errtion(
-				{
-					message: `There were no suitable editions for package [${opts.packagePath}]`,
-					code: 'unsupported-editions-tried',
-				},
-				editionsError
-			)
-		} else {
-			throw errtion(
-				{
-					message: 'There were no suitable editions',
-					code: 'unsupported-editions-tried',
-				},
-				editionsError
-			)
-		}
-	}
+	// this should never reach here
+	throw errtion({
+		message: `Unable to determine a suitable edition, as an unexpected pathway occured.`,
+		code: 'editions-autoloader-never',
+	})
 }
 
 /**
- * Cycle through the editions for a package and require the correct one.
- * Providing the package configuration is valid, editions handling is forwarded to {@link requireEditions}.
- * @returns The result of the loaded edition.
- * @throws An error if a suitable edition was unable to be resolved.
+ * Determine which edition should be loaded, and attempt to load it.
+ * @returns the loaded result of the suitable edition
+ * @throws if no suitable editions, or the edition failed to load
  */
-export function requirePackage(
-	cwd: EditionOptions['cwd'],
-	require: EditionOptions['require'],
-	entry: EditionOptions['entry']
-): any {
-	// Load the package.json file to fetch `name` for debugging and `editions` for loading
+export function solicitEdition<T>(editions: Editions, opts: SolicitOptions): T {
+	const edition = determineEdition(editions, opts)
+	return loadEdition<T>(edition, opts)
+}
+
+/**
+ * Cycle through the editions for a package, determine the compatible edition, and load it.
+ * @returns the loaded result of the suitable edition
+ * @throws if no suitable editions, or if the edition failed to load
+ */
+export function requirePackage<T>(
+	cwd: PathOptions['cwd'],
+	loader: LoaderOptions['loader'],
+	entry: PathOptions['entry']
+): T {
+	// load editions
 	const packagePath = resolve(cwd || '', 'package.json')
-	const { editions } = require(packagePath)
-	const opts = { packagePath, cwd, require, entry }
-	return requireEditions(editions, opts)
+	const { editions } = JSON.parse(readFileSync(packagePath, 'utf8'))
+	// load edition
+	return solicitEdition<T>(editions, {
+		versions: (processVersions as any) as Versions,
+		cwd,
+		loader,
+		entry,
+	})
 }