+++ source/test.ts
@@ -1,61 +1,41 @@
-/* eslint no-console:0 */
-
-// Blacklist
-process.env.EDITIONS_TAG_BLACKLIST = 'blacklist'
-
 // External
 import kava from 'kava'
 import { equal } from 'assert-helpers'
+import { versions as processVersions } from 'process'
 
 // Local
-import { Edition, requireEditions } from './index.js'
-import { simplifyRange } from './util.js'
-function pass() {
-	return 'ok'
-}
+import { Edition, solicitEdition, Editions } from './index.js'
 function fail() {
-	throw Error('not ok')
-}
-class PassThrough {
-	data: string
-	constructor() {
-		this.data = ''
-	}
-	write(data: any) {
-		this.data += data.toString()
-	}
+	throw Error('deliberate failure')
 }
 
 // Fixtures
 interface Fixture {
 	test: string
 	error?: string
-	strict?: boolean
-	require?: Function
-	result?: any
-	stderr?: string
-	blacklist?: string[]
-	editions: null | Array<Edition | {}>
+	loader?: Function
+	expected?: string
+	editions: null | Array<{}>
 }
 const fixtures: Fixture[] = [
 	{
 		test: 'missing editions',
-		error: 'no editions specified',
+		error: 'editions-autoloader-editions-missing',
 		editions: null,
 	},
 	{
 		test: 'missing editions',
-		error: 'no editions specified',
+		error: 'editions-autoloader-editions-missing',
 		editions: [],
 	},
 	{
 		test: 'missing all fields',
-		error: 'fields defined',
+		error: 'editions-autoloader-invalid-edition',
 		editions: [{}],
 	},
 	{
 		test: 'missing engines',
-		error: 'fields defined',
+		error: 'editions-autoloader-invalid-edition',
 		editions: [
 			{
 				description: 'value',
@@ -66,7 +46,7 @@ const fixtures: Fixture[] = [
 	},
 	{
 		test: 'missing entry',
-		error: 'fields defined',
+		error: 'editions-autoloader-invalid-edition',
 		editions: [
 			{
 				description: 'value',
@@ -77,7 +57,7 @@ const fixtures: Fixture[] = [
 	},
 	{
 		test: 'missing directory',
-		error: 'fields defined',
+		error: 'editions-autoloader-invalid-edition',
 		editions: [
 			{
 				description: 'value',
@@ -88,7 +68,7 @@ const fixtures: Fixture[] = [
 	},
 	{
 		test: 'missing description',
-		error: 'fields defined',
+		error: 'editions-autoloader-invalid-edition',
 		editions: [
 			{
 				description: 'value',
@@ -99,7 +79,7 @@ const fixtures: Fixture[] = [
 	},
 	{
 		test: 'skipped due to false engines',
-		error: 'its engines field was false',
+		error: 'editions-autoloader-invalid-engines',
 		editions: [
 			{
 				description: 'value',
@@ -110,8 +90,8 @@ const fixtures: Fixture[] = [
 		],
 	},
 	{
-		test: 'skipped due to falsey engines.node',
-		error: 'its .engines.node field was falsey',
+		test: 'skipped due to missing engines values',
+		error: 'editions-autoloader-engine-mismatch',
 		editions: [
 			{
 				description: 'value',
@@ -122,8 +102,8 @@ const fixtures: Fixture[] = [
 		],
 	},
 	{
-		test: 'skipped due to falsey engines.node',
-		error: 'its .engines.node field was falsey',
+		test: 'skipped due to engines.node = false',
+		error: 'editions-autoloader-engine-unsupported',
 		editions: [
 			{
 				description: 'value',
@@ -136,24 +116,8 @@ const fixtures: Fixture[] = [
 		],
 	},
 	{
-		test: 'skipped due to unsupported node version [strict=true]',
-		error: 'current node version',
-		strict: true,
-		editions: [
-			{
-				description: 'value',
-				entry: 'value',
-				directory: 'value',
-				engines: {
-					node: '0.6',
-				},
-			},
-		],
-	},
-	{
 		test: 'skipped due to unsupported node version',
-		error: 'it failed to load',
-		require: fail,
+		error: 'editions-autoloader-engine-incompatible',
 		editions: [
 			{
 				description: 'value',
@@ -166,28 +130,12 @@ const fixtures: Fixture[] = [
 		],
 	},
 	{
-		test: 'skipped due to blacklisted tag',
-		error: 'it contained a blacklisted tag',
-		blacklist: ['blacklist'],
+		test: 'skipped due to deliberate failure',
+		error: 'editions-autoloader-loader-failed',
+		loader: fail,
 		editions: [
 			{
-				description: 'value',
-				entry: 'value',
-				directory: 'value',
-				tags: ['blacklist'],
-				engines: {
-					node: '>=0.6',
-				},
-			},
-		],
-	},
-	{
-		test: 'skipped due to error',
-		error: 'it failed to load',
-		require: fail,
-		editions: [
-			{
-				description: 'value',
+				description: 'deliberate failure edition',
 				entry: 'value',
 				directory: 'value',
 				engines: {
@@ -199,11 +147,13 @@ const fixtures: Fixture[] = [
 	// pass test cases
 	{
 		test: 'loaded fine',
-		require: pass,
-		result: 'ok',
+		loader(this: Edition) {
+			return this.description
+		},
+		expected: 'is ok',
 		editions: [
 			{
-				description: 'value',
+				description: 'is ok',
 				entry: 'value',
 				directory: 'value',
 				engines: {
@@ -214,12 +164,13 @@ const fixtures: Fixture[] = [
 	},
 	{
 		test: 'loaded last edition',
-		require: pass,
-		result: 'ok',
-		stderr: 'current node version',
+		loader(this: Edition) {
+			return this.description
+		},
+		expected: `${process.versions.node} ok`,
 		editions: [
 			{
-				description: 'value',
+				description: 'is not ok',
 				entry: 'value',
 				directory: 'value',
 				engines: {
@@ -227,7 +178,7 @@ const fixtures: Fixture[] = [
 				},
 			},
 			{
-				description: 'value',
+				description: `${process.versions.node} ok`,
 				entry: 'value',
 				directory: 'value',
 				engines: {
@@ -240,34 +191,19 @@ const fixtures: Fixture[] = [
 
 // Test
 kava.suite('editions', function (suite, test) {
-	test('simplifyRange', function () {
-		equal(simplifyRange('4 || 6'), '>=4')
-		equal(simplifyRange('4'), '>=4')
-		equal(simplifyRange('4.0.0'), '>=4.0.0')
-		equal(simplifyRange('4.0.0-beta'), '>=4.0.0-beta')
-		equal(simplifyRange('4.0.0-beta || 5.0.0-beta'), '>=4.0.0-beta')
-	})
 	suite('requireEditions', function (suite, test) {
 		fixtures.forEach(function (fixture) {
 			test(fixture.test, function (done) {
 				try {
 					const opts = {
-						strict: fixture.strict,
-						require: fixture.require,
-						verbose: true,
-						stderr: new PassThrough(),
+						loader: fixture.loader,
+						versions: processVersions,
 					}
-					const result = requireEditions(
-						fixture.editions as Edition[],
+					const result = solicitEdition(
+						fixture.editions as Editions,
 						opts as any
 					)
-					equal(result, fixture.result, 'result was as expected')
-					if (fixture.stderr) {
-						const expected = opts.stderr.data.indexOf(fixture.stderr) !== -1
-						if (!expected) {
-							throw new Error('stderr was not as expected: ' + opts.stderr.data)
-						}
-					}
+					equal(result, fixture.expected, 'result was as expected')
 				} catch (err) {
 					if (fixture.error) {
 						const expected = err.stack.toString().indexOf(fixture.error) !== -1