+++ source/util.ts
@@ -1,17 +1,18 @@
-'use strict'
-
 // As Errlop uses Editions, we should use a specific Errlop edition
 // As otherwise, the circular reference may fail on some machines
 // https://github.com/bevry/errlop/issues/2
 import Errlop from 'errlop'
 
 interface ErrtionOptions {
+	/** The message to use for the error */
 	message: string
+	/** The code to categorise the error */
 	code: string | number
+	/** The severity level of the error */
 	level?: string | number
 }
 
-interface Errtion extends Errlop, ErrtionOptions {}
+export interface Errtion extends Errlop, ErrtionOptions {}
 
 /**
  * Allow code and level inputs on Errlop.
@@ -22,19 +23,12 @@ export function errtion(
 	opts: ErrtionOptions,
 	parent?: Errlop | Error
 ): Errtion {
-	const { message, code, level } = opts
+	const { code, level } = opts
+	let { message } = opts
+	if (code) message = `${code}: ${message}`
+	if (level) message = `${level}: ${message}`
 	const error = new Errlop(message, parent) as Errtion
 	if (code) error.code = code
 	if (level) error.level = level
 	return error
-}
-
-/** Converts anything to a string, by returning strings and serialising objects. */
-export function stringify(value: any): string {
-	return typeof value === 'string' ? value : JSON.stringify(value)
-}
-
-/** Converts a version range like `4 || 6` to `>=4` */
-export function simplifyRange(range: string): string {
-	return range.replace(/^([.\-\w]+)(\s.+)?$/, '>=$1')
 }